Editor: don't delete blueprint node when typing its classname

+++ editor/src/main/kotlin/hunternif/voxarch/editor/gui/GuiBlueprintEditor.kt
@@ -181,7 +181,9 @@ class GuiBlueprintEditor(
             }
         }
 
-        if (isEditorFocused && ImGui.isKeyPressed(GLFW.GLFW_KEY_DELETE, false)) {
+        if (isEditorFocused && !ImGui.isAnyItemActive() &&
+            ImGui.isKeyPressed(GLFW.GLFW_KEY_DELETE, false)
+        ) {
             deleteSelected()
         }
     }
